# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Constancia Backend API",
    "description": "Backend API contract for the Constancia frontend and Discord bot.",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/campaigns/{id}/events/{eventId}/fire": {
      "post": {
        "operationId": "fireEvent",
        "summary": "Fire an event pipeline",
        "tags": ["events"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "eventId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "eventId": {
                          "type": "string"
                        },
                        "messages": {
                          "type": "array",
                          "items": {
                            "anyOf": [
                              {
                                "type": "object",
                                "additionalProperties": false,
                                "properties": {
                                  "target": {
                                    "type": "string",
                                    "enum": ["player"]
                                  },
                                  "targetId": {
                                    "type": "string"
                                  },
                                  "content": {
                                    "type": "string"
                                  },
                                  "imageUrl": {
                                    "type": "string"
                                  }
                                },
                                "required": ["target", "content"]
                              },
                              {
                                "type": "object",
                                "additionalProperties": false,
                                "properties": {
                                  "target": {
                                    "type": "string",
                                    "enum": ["channel"]
                                  },
                                  "content": {
                                    "type": "string"
                                  },
                                  "imageUrl": {
                                    "type": "string"
                                  }
                                },
                                "required": ["target", "content"]
                              },
                              {
                                "type": "object",
                                "additionalProperties": false,
                                "properties": {
                                  "target": {
                                    "type": "string",
                                    "enum": ["group"]
                                  },
                                  "targetIds": {
                                    "type": "array",
                                    "items": {
                                      "type": "string"
                                    }
                                  },
                                  "content": {
                                    "type": "string"
                                  },
                                  "imageUrl": {
                                    "type": "string"
                                  }
                                },
                                "required": ["target", "targetIds", "content"]
                              }
                            ]
                          }
                        },
                        "halted": {
                          "type": "boolean"
                        }
                      },
                      "required": ["eventId", "messages", "halted"]
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}